@@ -1,57 +1,58 @@
 {
   "name": "api",
-  "version": "0.5.2",
+  "version": "0.6.0",
   "private": true,
   "scripts": {
     "start": "node ./bin/www",
     "build": "rm -rf dist/ && ./node_modules/.bin/tsc",
+    "lint": "./node_modules/.bin/tslint -p .",
     "cover": "./node_modules/.bin/nyc npm test",
-    "test": "./node_modules/.bin/mocha -t 10000"
+    "test": "./node_modules/.bin/mocha -t 10000 --exit"
   },
   "dependencies": {
-    "ajv": "^5.2.0",
-    "aws-sdk": "^2.81.0",
-    "bluebird": "^3.5.0",
-    "body-parser": "~1.17.1",
+    "ajv": "^6.5.4",
+    "aws-sdk": "^2.344.0",
+    "bluebird": "^3.5.2",
+    "body-parser": "~1.18.3",
     "cookie-parser": "~1.4.3",
-    "debug": "~2.6.3",
-    "express": "~4.15.2",
-    "express-jwt": "^5.3.0",
+    "debug": "~4.1.0",
+    "express": "~4.16.4",
+    "express-jwt": "^5.3.1",
     "express-jwt-authz": "^1.0.0",
     "fs-extra": "^3.0.1",
     "glob": "^7.1.2",
     "is-invalid-path": "^0.1.0",
     "jsonschema": "^1.1.1",
-    "lodash": "^4.17.4",
-    "mongoose": "^4.11.0",
-    "morgan": "~1.8.1",
-    "multer": "^1.3.0",
+    "lodash": "^4.17.11",
+    "mongoose": "^5.3.7",
+    "morgan": "~1.9.1",
+    "multer": "^1.4.1",
     "validator.ts": "^0.2.2"
   },
   "devDependencies": {
-    "@types/bluebird": "^3.5.8",
-    "@types/body-parser": "^1.16.4",
-    "@types/cookie-parser": "^1.3.30",
-    "@types/express": "^4.0.36",
-    "@types/express-jwt": "0.0.35",
-    "@types/fs-extra": "^3.0.3",
-    "@types/glob": "^5.0.30",
-    "@types/joi": "^10.4.0",
-    "@types/lodash": "^4.14.68",
-    "@types/mongoose": "^4.7.18",
-    "@types/morgan": "^1.7.32",
-    "@types/multer": "^1.3.2",
-    "chai": "^4.0.2",
+    "@types/bluebird": "^3.5.24",
+    "@types/body-parser": "^1.17.0",
+    "@types/cookie-parser": "^1.4.1",
+    "@types/express": "^4.16.0",
+    "@types/express-jwt": "0.0.40",
+    "@types/fs-extra": "^5.0.4",
+    "@types/glob": "^7.1.1",
+    "@types/joi": "^13.6.2",
+    "@types/lodash": "^4.14.117",
+    "@types/mongoose": "^5.2.20",
+    "@types/morgan": "^1.7.35",
+    "@types/multer": "^1.3.7",
+    "chai": "^4.2.0",
     "chai-as-promised": "^7.1.1",
     "clear-require": "^2.0.0",
-    "jsonwebtoken": "^7.4.1",
-    "mocha": "^3.4.2",
-    "nyc": "^11.0.3",
+    "jsonwebtoken": "^8.3.0",
+    "mocha": "^5.2.0",
+    "nyc": "^13.1.0",
     "source-map-support": "^0.4.15",
     "supertest": "^3.0.0",
-    "tslint": "^5.4.3",
-    "tslint-config-airbnb": "^5.1.2",
-    "typescript": "^2.4.1"
+    "tslint": "^5.11.0",
+    "tslint-config-airbnb": "^5.11.0",
+    "typescript": "^3.1.3"
   },
   "nyc": {
     "exclude": [
