@@ -10,9 +10,9 @@ export interface S3StorageOptions {
       accessKeyId: string,
       secretAccessKey: string,
     },
-    region: string
-  },
-  bucket: string,
+    region: string;
+  };
+  bucket: string;
 }
 
 export const config = {
@@ -35,7 +35,7 @@ export const config = {
   schemaRequired: !!process.env.SCHEMA_REQUIRED,
   jwtKey: process.env.JWT_KEY || 'v3rys3cr3tK3y',
   authz: {
-    enabled: process.env.JWT_ENABLED === '' ? true : process.env.JWT_ENABLED as any == true,
+    enabled: JSON.parse(process.env.JWT_ENABLED || 'true'),
     readScope: _.split(process.env.JWT_READ_SCOPE || 'filebank:read', ','),
     writeScope: _.split(process.env.JWT_WRITE_SCOPE || 'filebank:write', ','),
     deleteScope: _.split(process.env.JWT_DELETE_SCOPE || 'filebank:delete', ','),
