@@ -1,5 +1,5 @@
-import { directory, DirectoryModel } from './directory';
-import { file, FileModel } from './file';
+import { directory } from './directory';
+import { file } from './file';
 // import * as _ from 'lodash';
 
 export default {
