@@ -2,9 +2,8 @@ import fs = require('fs');
 import path = require('path');
 import fsExtra = require('fs-extra');
 import * as glob from 'glob';
-import { FileStorage, Directory, File, ItemType, Item } from '../index';
-import * as Promise from 'bluebird';
-import { config, FilesystemStorageOptions } from '../config';
+import { Directory, File, FileStorage, Item, ItemType } from '../index';
+import * as Bluebird from 'bluebird';
 
 const DEFAULT_MIMETYPE = 'application/octet-stream';
 
@@ -36,42 +35,38 @@ export class FilesystemStorage implements FileStorage {
     }
     return path;
   }
-  list(id: string): Promise<(Directory | File)[]> {
+  async list(id: string): Promise<(Directory | File)[]> {
     const dirPath : string = path.resolve(this.rootDir, id);
-    return Promise.fromNode(cb => fs.readdir(dirPath, cb))
-      .then((filenames: string[]) => {
-        const items: (Directory | File)[] = [];
-        filenames.map((filename) => {
-          const ref = this.relatify(path.resolve(this.rootDir, id, filename));
-          if (fs.statSync(path.resolve(dirPath, filename)).isDirectory()) {
-            items.push({
-              refId: ref,
-              name: filename,
-              type: ItemType.Directory,
-            });
-          } else {
-            items.push({
-              refId: ref,
-              name: filename,
-              mimetype: DEFAULT_MIMETYPE,
-              type: ItemType.File,
-            });
-          }
+    const filenames: string[] = await Bluebird.fromNode<string[]>(cb => fs.readdir(dirPath, cb));
+    const items: (Directory | File)[] = [];
+    filenames.map((filename) => {
+      const ref = this.relatify(path.resolve(this.rootDir, id, filename));
+      if (fs.statSync(path.resolve(dirPath, filename)).isDirectory()) {
+        items.push({
+          refId: ref,
+          name: filename,
+          type: ItemType.Directory,
         });
-        return items;
-      });
+      } else {
+        items.push({
+          refId: ref,
+          name: filename,
+          mimetype: DEFAULT_MIMETYPE,
+          type: ItemType.File,
+        });
+      }
+    });
+    return items;
   }
 
-  mkdir(dirRef: string, dir: Directory): Promise<Directory> {
+  async mkdir(dirRef: string, dir: Directory): Promise<Directory> {
     const dirPath = path.resolve(this.rootDir, dirRef || '', dir.name);
-    return Promise.fromCallback(cb => fs.mkdir(dirPath, cb))
-      .then(() => {
-        return {
-          refId: this.relatify(dirPath),
-          name: dir.name,
-          type: ItemType.Directory,
-        };
-      });
+    await Bluebird.fromNode(cb => fs.mkdir(dirPath, cb));
+    return {
+      refId: this.relatify(dirPath),
+      name: dir.name,
+      type: ItemType.Directory,
+    };
   }
 
   rmdir(dirId: string): Promise<void> {
@@ -83,28 +78,26 @@ export class FilesystemStorage implements FileStorage {
     }
   }
 
-  getContent(file: File): Promise<Buffer> {
-    return Promise.fromCallback(cb => fs.readFile(path.resolve(this.rootDir, file.refId), cb));
+  async getContent(file: File): Promise<Buffer> {
+    return Bluebird.fromNode<Buffer>(cb => fs.readFile(path.resolve(this.rootDir, file.refId), cb));
   }
 
-  removeFile(file: File): Promise<void> {
-    return Promise.fromCallback(cb => fs.unlink(path.resolve(this.rootDir, file.refId), cb));
+  async removeFile(file: File): Promise<void> {
+    await Bluebird.fromNode(cb => fs.unlink(path.resolve(this.rootDir, file.refId), cb));
   }
 
-  createFile(file: File, directory: Directory|undefined, data: Buffer): Promise<File> {
+  async createFile(file: File, directory: Directory|undefined, data: Buffer): Promise<File> {
     const filePath = path.resolve(this.rootDir, directory ? directory.refId : '', file.name);
-    return Promise.fromCallback(cb => fs.writeFile(filePath, data, cb))
-      .then(() => {
-        return {
-          mimetype: DEFAULT_MIMETYPE,
-          refId: this.relatify(filePath),
-          name: file.name,
-          type: ItemType.File,
-        };
-      });
+    await Bluebird.fromCallback(cb => fs.writeFile(filePath, data, cb));
+    return {
+      mimetype: DEFAULT_MIMETYPE,
+      refId: this.relatify(filePath),
+      name: file.name,
+      type: ItemType.File,
+    };
   }
 
-  moveFile(file: File, destination: Directory): Promise<File> {
+  async moveFile(file: File, destination: Directory): Promise<File> {
     function move(oldPath: string, newPath: string, callback: (err?: NodeJS.ErrnoException) => void) {
       fs.rename(oldPath, newPath, (err) => {
         if (err) {
@@ -134,38 +127,32 @@ export class FilesystemStorage implements FileStorage {
 
     const oldPath = path.resolve(this.rootDir, file.refId);
     const filename = file.refId.split('/').pop();
-    const destPath = path.resolve(this.rootDir, destination.refId, filename);
-    return Promise.fromCallback(cb => move(oldPath, destPath, cb))
-      .then(() => {
-        file.refId = this.relatify(destPath);
-        return file;
-      });
+    const destPath = path.resolve(this.rootDir, destination.refId, filename || '');
+    await Bluebird.fromCallback(cb => move(oldPath, destPath, cb));
+    file.refId = this.relatify(destPath);
+    return file;
   }
 
-  moveDirectory(dir: Directory, destination: Directory, refCb: (oldItem: Item, newItem: Item) => Promise<void>): Promise<Directory> {
+  async moveDirectory(dir: Directory, destination: Directory, refCb: (oldItem: Item, newItem: Item) => Promise<void>): Promise<Directory> {
     const subdirRoot = path.resolve(this.rootDir, dir.refId);
-    return Promise.fromNode(cb => glob('**', { cwd: subdirRoot }, cb))
-      .then((items) => {
-        const destinationRef = destination.refId ? destination.refId : destination.name;
-        const targetPath = fs.existsSync(path.resolve(this.rootDir, destinationRef)) ?
-          path.resolve(this.rootDir, destinationRef, path.basename(dir.refId)) :
-          path.resolve(this.rootDir, destinationRef);
+    const items = await Bluebird.fromNode<string[]>(cb => glob('**', { cwd: subdirRoot }, cb));
+    const destinationRef = destination.refId ? destination.refId : destination.name;
+    const targetPath = fs.existsSync(path.resolve(this.rootDir, destinationRef)) ?
+      path.resolve(this.rootDir, destinationRef, path.basename(dir.refId)) :
+      path.resolve(this.rootDir, destinationRef);
 
-        fsExtra.moveSync(subdirRoot, targetPath);
-        const subitemPromises = items.map((item: string) => {
-          const stat = fs.statSync(path.resolve(targetPath, item));
-          const oldItem = { name: path.basename(item), refId: path.join(dir.refId, item), type: stat.isDirectory() ? ItemType.Directory : ItemType.File };
-          const newItem = {
-            name: path.basename(item),
-            refId: path.join(destinationRef, item),
-            type: stat.isDirectory() ? ItemType.Directory : ItemType.File,
-          };
-          return refCb(oldItem, newItem);
-        });
-        return Promise.all(subitemPromises).then(() => {
-          return { type: ItemType.Directory, name: path.basename(targetPath), refId: this.relatify(targetPath) };
-        });
-      });
+    fsExtra.moveSync(subdirRoot, targetPath);
+    items.map((item: string) => {
+      const stat = fs.statSync(path.resolve(targetPath, item));
+      const oldItem = { name: path.basename(item), refId: path.join(dir.refId, item), type: stat.isDirectory() ? ItemType.Directory : ItemType.File };
+      const newItem = {
+        name: path.basename(item),
+        refId: path.join(destinationRef, item),
+        type: stat.isDirectory() ? ItemType.Directory : ItemType.File,
+      };
+      return refCb(oldItem, newItem);
+    });
+    return { type: ItemType.Directory, name: path.basename(targetPath), refId: this.relatify(targetPath) };
   }
 
   setRootDir(dir: string) {
